requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
numpy
scikit-learn
tqdm
pandas
python-dotenv
gpt4all
pypdf2
